feat: babel-preset-ice
