+++ packages/build-scripts-config/src/config/babel/index.js
@@ -1,52 +1,13 @@
-/**
- * babel config
- */
-function resolvePlugin(plugins) {
-  return plugins.filter(Boolean).map((plugin) => {
-    if (Array.isArray(plugin)) {
-      const [pluginName, ...args] = plugin;
-      return [require.resolve(pluginName), ...args];
-    }
-    return require.resolve(plugin);
-  });
-}
+const getBabelPreset = require('babel-preset-ice').default;
 
 module.exports = () => {
-  const plugins = [
-    // Stage 0
-    '@babel/plugin-proposal-function-bind',
-    // Stage 1
-    '@babel/plugin-proposal-export-default-from',
-    '@babel/plugin-proposal-logical-assignment-operators',
-    ['@babel/plugin-proposal-optional-chaining', { loose: false }],
-    ['@babel/plugin-proposal-pipeline-operator', { proposal: 'minimal' }],
-    ['@babel/plugin-proposal-nullish-coalescing-operator', { loose: false }],
-    '@babel/plugin-proposal-do-expressions',
-    // Stage 2
-    ['@babel/plugin-proposal-decorators', { legacy: true }],
-    '@babel/plugin-proposal-function-sent',
-    '@babel/plugin-proposal-export-namespace-from',
-    '@babel/plugin-proposal-numeric-separator',
-    '@babel/plugin-proposal-throw-expressions',
-    // Stage 3
-    '@babel/plugin-syntax-dynamic-import',
-    '@babel/plugin-syntax-import-meta',
-    ['@babel/plugin-proposal-class-properties', { loose: true }],
-    '@babel/plugin-proposal-json-strings',
-  ];
-
-  return {
-    presets: resolvePlugin([
-      [
-        '@babel/preset-env', {
-          modules: false,
-          useBuiltIns: 'entry',
-          corejs: 3,
-        },
-      ],
-      '@babel/preset-typescript',
-      '@babel/preset-react',
-    ]),
-    plugins: resolvePlugin(plugins),
-  };
+  return getBabelPreset({
+    react: true,
+    typescript: true,
+    env: {
+      modules: false,
+      useBuiltIns: 'entry',
+      corejs: 3,
+    }
+  });
 };
